[refactor] #17 LoginRequestDto 삭제 및 Dto 패키지 재정립
