--- src/main/java/com/wedit/weditapp/domain/member/dto/LoginRequestDto.java
@@ -1,12 +1,0 @@
-package com.wedit.weditapp.domain.member.dto;
-
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Getter
-@NoArgsConstructor
-public class LoginRequestDto {
-    private String email;
-
-}
